refactor: Restructure home feature packages by moving dialogs and debug panel into dedicated packages
