+++ feature/home/src/main/java/com/scrolless/app/feature/home/HomeScreen.kt
@@ -128,13 +128,13 @@ import com.scrolless.app.designsystem.util.formatTime
 import com.scrolless.app.designsystem.util.radialGradientScrim
 import com.scrolless.app.designsystem.util.toCountdownLabel
 import com.scrolless.app.designsystem.util.toIntervalLabel
-import com.scrolless.app.feature.home.components.AccessibilityExplainerBottomSheet
-import com.scrolless.app.feature.home.components.AccessibilitySuccessBottomSheet
-import com.scrolless.app.feature.home.components.AccessibilitySuccessBottomSheetPreview
-import com.scrolless.app.feature.home.components.FloatingDebugUsagePanel
-import com.scrolless.app.feature.home.components.HelpDialog
-import com.scrolless.app.feature.home.components.IntervalTimerDialog
-import com.scrolless.app.feature.home.components.TimeLimitDialog
+import com.scrolless.app.feature.home.debug.FloatingDebugUsagePanel
+import com.scrolless.app.feature.home.dialogs.AccessibilityExplainerBottomSheet
+import com.scrolless.app.feature.home.dialogs.AccessibilitySuccessBottomSheet
+import com.scrolless.app.feature.home.dialogs.AccessibilitySuccessBottomSheetPreview
+import com.scrolless.app.feature.home.dialogs.HelpDialog
+import com.scrolless.app.feature.home.dialogs.IntervalTimerDialog
+import com.scrolless.app.feature.home.dialogs.TimeLimitDialog
 import java.time.LocalDateTime
 import java.util.concurrent.TimeUnit
 import kotlinx.coroutines.delay
