+++ feature/home/src/main/java/com/scrolless/app/feature/home/dialogs/AccessibilityExplainerDialog.kt
@@ -14,7 +14,7 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
-package com.scrolless.app.feature.home.components
+package com.scrolless.app.feature.home.dialogs
 
 import android.content.Intent
 import androidx.compose.animation.AnimatedVisibility
